refactor : 포스트 생성 일 및 지난 시간 추출 함수 추상화 및 분리
