+++ src/pages/Home/components/post-detail/PostHeader.tsx
@@ -1,3 +1,5 @@
+import getPostLiveTime from "@/utils/getPostCreateTime";
+
 type PostHeaderType = {
     title : string,
     author : string,
@@ -19,8 +21,8 @@ const PostHeader = ({title, author, createAt, profileImg} : PostHeaderType) =>{
                 <p className ="font-semibold text-[1.1rem]">{author}</p>
                 <p className ="text-[0.65rem] my-auto text-purple-500 font-medium">팔로잉 중</p>
             </div>
-            <p className ="w-fit mx-auto text-[1.4rem] font-medium my-1">{title}</p>
-            <p className ="my-3">{createAt}</p>
+            <p className ="w-fit mx-auto text-[1.4rem] font-medium my-2">{title}</p>
+            <p className ="my-2 w-fit mx-auto text-[0.8rem]">{getPostLiveTime(createAt)}</p>
             <div className ="flex w-fit mx-auto gap-3 my-3">
                 <button className ="border-slate-200 border-1 px-3 py-1 rounded-lg bg-[#efb98b]">
                     수정
